+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 3.2.0
+
+### Changed
+- The Windows clipboard now behaves consistently with the other
+platform implementations again.
+
 ## 3.1.1 on 2022-17-10
 
 ### Added